Collection of code fixes and improvements
Resurected itek plot files that were broken by inconsistent variable
passing between subroutines.  This is now part of the rfile tests.
The content mostly matches what it should, but there are a few glitches
left to iron out.  Global variable sizes are now set explicitly rather
than being passed implicitly and all integer values are explicitly set
to 4 bytes.  If 8 bytes are needed somewhere the changes should be made
more carefully.

get_expaths and efit_version files were merged into a new subroutine
called set_exvars that is configured by CMAKE to set all external
variables before compilation.

Test execution scripts set OMP_NUM_THREADS to 1 to avoid unwanted
threading that was seen on some systems (even though there is no
OpenMP code directly in efit yet).

A few fixes were made to Iris configure files and documentation.

Several bugs were fixed, files were named more consistently, unused
functions and variables were removed, and formatting was improved.

+++ efit/autoknot.F90
@@ -15,7 +15,7 @@
       implicit integer*4 (i-n), real*8 (a-h,o-z)
 
       external ppakfunc,ffakfunc,wwakfunc,eeakfunc
-      integer, intent(inout) :: kerror
+      integer*4, intent(inout) :: kerror
       kerror = 0
 
 !
